feat: Allow defining binding class types using kotlin function syntax refactor: Delete Table, View, etc... classes

+++ src/commonMain/kotlin/me/dvyy/sqlite/Identity.kt
@@ -1,3 +1,3 @@
 package me.dvyy.sqlite
 
-typealias Identity = Int
+typealias Identity = Long
